Refactor program getKey test
Refactors the program getKey test to use a new AppEngineTest base
class.

After this commit, new test classes should inherit from AppEngineTest
when they require the use of App Engine SDKs, like the
PersistenceManager.
